mental health common component
